chore: rename Events to Logbook
