+++ examples/asm/get_host_risks.py
@@ -1,11 +1,11 @@
 """Retrieve host risks and subsequent host info for each host risk."""
-from censys.asm import Events, HostsAssets
+from censys.asm import HostsAssets, Logbook
 
-e = Events()
-h = HostsAssets()
+logbook = Logbook()
+hosts = HostsAssets()
 
-cursor = e.get_cursor(filters=["HOST_RISK"])
-events = e.get_events(cursor)
+cursor = logbook.get_cursor(filogbookters=["HOST_RISK"])
+events = logbook.get_events(cursor)
 
 for event in events:
     # only show logbook events with the 'add' tag
@@ -13,5 +13,5 @@ for event in events:
         print(event)
 
         # enrich the data of this host_risk with more data from the host itself
-        host = h.get_asset_by_id(event["entity"]["ipAddress"])
+        host = hosts.get_asset_by_id(event["entity"]["ipAddress"])
         print(host)
